requirements.txt
streamlit
langchain
python-dotenv
unstructured
langchain-google-genai
faiss-cpu
langchain-community